Componentize application architecture
Split monolithic text_collector.py into modular components:
- core.py: Main application logic
- cli.py: Command line interface
- chromadb.py: Database operations
- chunker.py: Text chunking functionality
- claude.py/deepseek.py: API clients
- file_io.py: File operations
- __main__.py: Entry point
