+++ .gitignore
@@ -1,2 +1,3 @@
 env/
 chroma_db/
+__pycache__/
