+++ __init__.py
@@ -1,0 +1,5 @@
+"""
+TextCollector - Semantic search and question answering for your text collections.
+"""
+
+__version__ = "0.1.0"
