+++ __main__.py
@@ -1,0 +1,4 @@
+from TextCollector.cli import main
+
+if __name__ == "__main__":
+    exit(main())
